+++ packages/react-hooks/src/index.ts
@@ -1,3 +1,8 @@
 export { default as usePrevious } from './usePrevious';
 export { default as useForwardedRef } from './useForwardedRef';
 export { default as useCopyToClipboard } from './useCopyToClipboard';
+export { default as usePromiseFactory } from './usePromiseFactory';
+export type {
+  UsePromiseFactoryOptions,
+  UsePromiseFactoryResult,
+} from './usePromiseFactory';